refactor(studio-api): split server routes and config helpers
