+++ packages/studio/src/api/routes/books.ts
@@ -1,0 +1,228 @@
+import type { Hono } from "hono";
+import {
+  createInteractionToolsFromDeps,
+  PipelineRunner,
+  processProjectInteractionRequest,
+  type PipelineConfig,
+  type StateManager,
+} from "@actalk/inkos-core";
+import { access, readFile, readdir, writeFile } from "node:fs/promises";
+import { join } from "node:path";
+import { buildStudioBookConfig } from "../book-create.js";
+import { loadStudioBookListSummary } from "../books/summary.js";
+
+interface RegisterBookRoutesOptions {
+  readonly root: string;
+  readonly state: StateManager;
+  readonly bookCreateStatus: Map<string, { status: "creating" | "error"; error?: string }>;
+  readonly broadcast: (event: string, data: unknown) => void;
+  readonly buildPipelineConfig: () => Promise<PipelineConfig>;
+}
+
+export function registerBookRoutes(app: Hono, options: RegisterBookRoutesOptions): void {
+  app.get("/api/v1/books", async (c) => {
+    const bookIds = await options.state.listBooks();
+    const books = await Promise.all(bookIds.map((id) => loadStudioBookListSummary(options.state, id)));
+    return c.json({ books });
+  });
+
+  app.get("/api/v1/books/:id", async (c) => {
+    const id = c.req.param("id");
+    try {
+      const book = await options.state.loadBookConfig(id);
+      const chapters = await options.state.loadChapterIndex(id);
+      const nextChapter = await options.state.getNextChapterNumber(id);
+      return c.json({ book, chapters, nextChapter });
+    } catch {
+      return c.json({ error: `Book "${id}" not found` }, 404);
+    }
+  });
+
+  app.post("/api/v1/books/create", async (c) => {
+    const body = await c.req.json<{
+      title: string;
+      genre: string;
+      language?: string;
+      platform?: string;
+      chapterWordCount?: number;
+      targetChapters?: number;
+      blurb?: string;
+    }>();
+
+    const now = new Date().toISOString();
+    const bookConfig = buildStudioBookConfig(body, now);
+    const bookId = bookConfig.id;
+    const bookDir = options.state.bookDir(bookId);
+
+    try {
+      await access(join(bookDir, "book.json"));
+      await access(join(bookDir, "story", "story_bible.md"));
+      return c.json({ error: `Book "${bookId}" already exists` }, 409);
+    } catch {
+      // The target book is not fully initialized yet, so creation can continue.
+    }
+
+    options.broadcast("book:creating", { bookId, title: body.title });
+    options.bookCreateStatus.set(bookId, { status: "creating" });
+
+    const pipeline = new PipelineRunner(await options.buildPipelineConfig());
+    const tools = createInteractionToolsFromDeps(pipeline, options.state);
+    processProjectInteractionRequest({
+      projectRoot: options.root,
+      request: {
+        intent: "create_book",
+        title: body.title,
+        genre: body.genre,
+        language: body.language === "en" ? "en" : body.language === "zh" ? "zh" : undefined,
+        platform: body.platform,
+        chapterWordCount: body.chapterWordCount,
+        targetChapters: body.targetChapters,
+        blurb: body.blurb,
+      },
+      tools,
+    }).then(
+      async (result: {
+        readonly session: { readonly activeBookId?: string };
+        readonly details?: Readonly<Record<string, unknown>>;
+      }) => {
+        const createdBookId = (result.details?.bookId as string | undefined) ?? result.session.activeBookId ?? bookId;
+        const book = await loadStudioBookListSummary(options.state, createdBookId).catch(() => undefined);
+        options.bookCreateStatus.delete(createdBookId);
+        options.broadcast("book:created", { bookId: createdBookId, ...(book ? { book } : {}) });
+      },
+      (e: unknown) => {
+        const error = e instanceof Error ? e.message : String(e);
+        options.bookCreateStatus.set(bookId, { status: "error", error });
+        options.broadcast("book:error", { bookId, error });
+      },
+    );
+
+    return c.json({ status: "creating", bookId });
+  });
+
+  app.get("/api/v1/books/:id/create-status", async (c) => {
+    const id = c.req.param("id");
+    const status = options.bookCreateStatus.get(id);
+    if (!status) {
+      return c.json({ status: "missing" }, 404);
+    }
+    return c.json(status);
+  });
+
+  app.get("/api/v1/books/:id/chapters/:num", async (c) => {
+    const id = c.req.param("id");
+    const num = parseInt(c.req.param("num"), 10);
+    const bookDir = options.state.bookDir(id);
+    const chaptersDir = join(bookDir, "chapters");
+
+    try {
+      const files = await readdir(chaptersDir);
+      const paddedNum = String(num).padStart(4, "0");
+      const match = files.find((f) => f.startsWith(paddedNum) && f.endsWith(".md"));
+      if (!match) return c.json({ error: "Chapter not found" }, 404);
+      const content = await readFile(join(chaptersDir, match), "utf-8");
+      return c.json({ chapterNumber: num, filename: match, content });
+    } catch {
+      return c.json({ error: "Chapter not found" }, 404);
+    }
+  });
+
+  app.put("/api/v1/books/:id/chapters/:num", async (c) => {
+    const id = c.req.param("id");
+    const num = parseInt(c.req.param("num"), 10);
+    const bookDir = options.state.bookDir(id);
+    const chaptersDir = join(bookDir, "chapters");
+    const { content } = await c.req.json<{ content: string }>();
+
+    try {
+      const files = await readdir(chaptersDir);
+      const paddedNum = String(num).padStart(4, "0");
+      const match = files.find((f) => f.startsWith(paddedNum) && f.endsWith(".md"));
+      if (!match) return c.json({ error: "Chapter not found" }, 404);
+
+      await writeFile(join(chaptersDir, match), content, "utf-8");
+      return c.json({ ok: true, chapterNumber: num });
+    } catch (e) {
+      return c.json({ error: String(e) }, 500);
+    }
+  });
+
+  app.post("/api/v1/books/:id/write-next", async (c) => {
+    const id = c.req.param("id");
+    const body = await c.req.json<{ wordCount?: number }>().catch(() => ({ wordCount: undefined }));
+
+    options.broadcast("write:start", { bookId: id });
+
+    const pipeline = new PipelineRunner(await options.buildPipelineConfig());
+    pipeline.writeNextChapter(id, body.wordCount).then(
+      (result) => {
+        options.broadcast("write:complete", { bookId: id, chapterNumber: result.chapterNumber, status: result.status, title: result.title, wordCount: result.wordCount });
+      },
+      (e) => {
+        options.broadcast("write:error", { bookId: id, error: e instanceof Error ? e.message : String(e) });
+      },
+    );
+
+    return c.json({ status: "writing", bookId: id });
+  });
+
+  app.post("/api/v1/books/:id/draft", async (c) => {
+    const id = c.req.param("id");
+    const body = await c.req.json<{ wordCount?: number; context?: string }>().catch(() => ({ wordCount: undefined, context: undefined }));
+
+    options.broadcast("draft:start", { bookId: id });
+
+    const pipeline = new PipelineRunner(await options.buildPipelineConfig());
+    pipeline.writeDraft(id, body.context, body.wordCount).then(
+      (result) => {
+        options.broadcast("draft:complete", { bookId: id, chapterNumber: result.chapterNumber, title: result.title, wordCount: result.wordCount });
+      },
+      (e) => {
+        options.broadcast("draft:error", { bookId: id, error: e instanceof Error ? e.message : String(e) });
+      },
+    );
+
+    return c.json({ status: "drafting", bookId: id });
+  });
+
+  app.post("/api/v1/books/:id/chapters/:num/approve", async (c) => {
+    const id = c.req.param("id");
+    const num = parseInt(c.req.param("num"), 10);
+
+    try {
+      const index = await options.state.loadChapterIndex(id);
+      const updated = index.map((ch) =>
+        ch.number === num ? { ...ch, status: "approved" as const } : ch,
+      );
+      await options.state.saveChapterIndex(id, updated);
+      return c.json({ ok: true, chapterNumber: num, status: "approved" });
+    } catch (e) {
+      return c.json({ error: String(e) }, 500);
+    }
+  });
+
+  app.post("/api/v1/books/:id/chapters/:num/reject", async (c) => {
+    const id = c.req.param("id");
+    const num = parseInt(c.req.param("num"), 10);
+
+    try {
+      const index = await options.state.loadChapterIndex(id);
+      const target = index.find((ch) => ch.number === num);
+      if (!target) {
+        return c.json({ error: `Chapter ${num} not found` }, 404);
+      }
+
+      const rollbackTarget = num - 1;
+      const discarded = await options.state.rollbackToChapter(id, rollbackTarget);
+      return c.json({
+        ok: true,
+        chapterNumber: num,
+        status: "rejected",
+        rolledBackTo: rollbackTarget,
+        discarded,
+      });
+    } catch (e) {
+      return c.json({ error: String(e) }, 500);
+    }
+  });
+}
