+++ packages/studio/src/api/routes/services.ts
@@ -1,0 +1,286 @@
+import type { Hono } from "hono";
+import {
+  getAllEndpoints,
+  isApiKeyOptionalForEndpoint,
+  listModelsForService,
+  loadSecrets,
+  probeModelsFromUpstream,
+  resolveServiceProviderFamily,
+  saveSecrets,
+} from "@actalk/inkos-core";
+import {
+  filterTextChatModels,
+  isCustomServiceId,
+  isTextChatModelId,
+  loadRawConfig,
+  mergeServiceConfig,
+  normalizeConfigSource,
+  normalizeServiceConfig,
+  probeServiceCapabilities,
+  readEnvConfigStatus,
+  resolveConfiguredServiceBaseUrl,
+  saveRawConfig,
+  type LLMConfigSource,
+} from "../services/service-config.js";
+
+interface RegisterServiceRoutesOptions {
+  readonly root: string;
+}
+
+const modelListCache = new Map<string, { models: Array<{ id: string; name: string }>; at: number }>();
+
+export function registerServiceRoutes(app: Hono, options: RegisterServiceRoutesOptions): void {
+  app.get("/api/v1/services", async (c) => {
+    const secrets = await loadSecrets(options.root);
+    const endpoints = getAllEndpoints().filter((ep) => ep.id !== "custom");
+
+    // Fast: only check connection status from secrets, no external API calls.
+    const services = endpoints.map((ep) => ({
+      service: ep.id,
+      label: ep.label,
+      group: ep.group,
+      connected: Boolean(secrets.services[ep.id]?.apiKey),
+    }));
+
+    // Add custom services from inkos.json
+    try {
+      const config = await loadRawConfig(options.root);
+      for (const svc of normalizeServiceConfig((config.llm as Record<string, unknown> | undefined)?.services)) {
+        if (svc.service === "custom") {
+          const secretKey = `custom:${svc.name}`;
+          services.push({
+            service: secretKey,
+            label: svc.name ?? "Custom",
+            group: undefined,
+            connected: Boolean(secrets.services[secretKey]?.apiKey),
+          });
+        }
+      }
+    } catch {
+      // no config file
+    }
+
+    return c.json({ services });
+  });
+
+  app.get("/api/v1/services/config", async (c) => {
+    const config = await loadRawConfig(options.root);
+    const llm = (config.llm as Record<string, unknown> | undefined) ?? {};
+    const services = normalizeServiceConfig(llm.services);
+    const envConfig = await readEnvConfigStatus(options.root);
+    return c.json({
+      services,
+      service: typeof llm.service === "string" ? llm.service : null,
+      defaultModel: llm.defaultModel ?? null,
+      configSource: "studio" satisfies LLMConfigSource,
+      storedConfigSource: normalizeConfigSource(llm.configSource),
+      envConfig,
+    });
+  });
+
+  app.put("/api/v1/services/config", async (c) => {
+    const body = await c.req.json<{ services?: unknown; defaultModel?: string; configSource?: LLMConfigSource; service?: string }>();
+    const config = await loadRawConfig(options.root);
+    config.llm = config.llm ?? {};
+    const llm = config.llm as Record<string, unknown>;
+    if (body.services !== undefined) {
+      const existingServices = normalizeServiceConfig(llm.services);
+      const incomingServices = normalizeServiceConfig(body.services);
+      llm.services = mergeServiceConfig(existingServices, incomingServices);
+    }
+    if (body.defaultModel !== undefined) {
+      llm.defaultModel = body.defaultModel;
+    }
+    if (body.configSource === "env") {
+      return c.json({
+        error: "Studio 运行时不支持切换到 env；env 只在 CLI/daemon/部署运行时作为覆盖层使用。",
+      }, 400);
+    }
+    if (body.configSource !== undefined) {
+      llm.configSource = normalizeConfigSource(body.configSource);
+    }
+    if (body.service !== undefined) {
+      llm.service = body.service;
+    }
+    await saveRawConfig(options.root, config);
+    return c.json({ ok: true });
+  });
+
+  app.post("/api/v1/services/:service/test", async (c) => {
+    const service = c.req.param("service");
+    const { apiKey, baseUrl, apiFormat, stream } = await c.req.json<{
+      apiKey: string;
+      baseUrl?: string;
+      apiFormat?: "chat" | "responses";
+      stream?: boolean;
+    }>();
+
+    const resolvedBaseUrl = await resolveConfiguredServiceBaseUrl(options.root, service, baseUrl);
+    if (!resolvedBaseUrl) {
+      return c.json({ ok: false, error: `未知服务商: ${service}` }, 400);
+    }
+
+    const baseService = isCustomServiceId(service) ? "custom" : service;
+    const apiKeyOptional = isApiKeyOptionalForEndpoint({
+      provider: resolveServiceProviderFamily(baseService) ?? "openai",
+      baseUrl: resolvedBaseUrl,
+    });
+    if (!apiKey?.trim() && !apiKeyOptional) {
+      return c.json({ ok: false, error: "API Key 不能为空" }, 400);
+    }
+
+    const rawConfig = await loadRawConfig(options.root).catch(() => ({} as Record<string, unknown>));
+    const llm = (rawConfig.llm as Record<string, unknown> | undefined) ?? {};
+    const probe = await probeServiceCapabilities({
+      root: options.root,
+      service,
+      apiKey: apiKey?.trim() ?? "",
+      baseUrl: resolvedBaseUrl,
+      preferredApiFormat: apiFormat,
+      preferredStream: stream,
+      proxyUrl: typeof llm.proxyUrl === "string" ? llm.proxyUrl : undefined,
+    });
+
+    const probeStatus = {
+      ok: probe.ok,
+      models: probe.models?.length ?? 0,
+      ...(probe.ok ? {} : { error: probe.error ?? "连接失败" }),
+    };
+
+    if (!probe.ok) {
+      return c.json({
+        ok: false,
+        error: probe.error ?? "连接失败",
+        probe: probeStatus,
+        chat: null,
+      }, 400);
+    }
+
+    return c.json({
+      ok: true,
+      modelCount: probe.models.length,
+      models: probe.models,
+      selectedModel: probe.selectedModel,
+      detected: {
+        apiFormat: probe.apiFormat,
+        stream: probe.stream,
+        baseUrl: probe.baseUrl,
+        modelsSource: probe.modelsSource,
+      },
+      probe: probeStatus,
+      chat: null,
+    });
+  });
+
+  app.put("/api/v1/services/:service/secret", async (c) => {
+    const service = c.req.param("service");
+    const { apiKey } = await c.req.json<{ apiKey: string }>();
+    const secrets = await loadSecrets(options.root);
+    if (apiKey?.trim()) {
+      secrets.services[service] = { apiKey: apiKey.trim() };
+    } else {
+      delete secrets.services[service];
+    }
+    await saveSecrets(options.root, secrets);
+    return c.json({ ok: true });
+  });
+
+  app.get("/api/v1/services/:service/secret", async (c) => {
+    const service = c.req.param("service");
+    const secrets = await loadSecrets(options.root);
+    return c.json({
+      apiKey: secrets.services[service]?.apiKey ?? "",
+    });
+  });
+
+  app.get("/api/v1/services/models", async (c) => {
+    const secrets = await loadSecrets(options.root);
+    const endpoints = getAllEndpoints()
+      .filter((ep) => ep.id !== "custom" && Boolean(secrets.services[ep.id]?.apiKey));
+
+    const groups = endpoints.map((ep) => ({
+      service: ep.id,
+      label: ep.label,
+      models: ep.models
+        .filter((m) => m.enabled !== false)
+        .filter((m) => isTextChatModelId(m.id))
+        .map((m) => ({
+          id: m.id,
+          name: m.id,
+          ...(typeof m.maxOutput === "number" ? { maxOutput: m.maxOutput } : {}),
+          ...(m.contextWindowTokens > 0 ? { contextWindow: m.contextWindowTokens } : {}),
+        })),
+    }));
+
+    return c.json({ groups });
+  });
+
+  app.get("/api/v1/services/models/custom", async (c) => {
+    const secrets = await loadSecrets(options.root);
+    let config: Record<string, unknown> = {};
+    try {
+      config = await loadRawConfig(options.root);
+    } catch {
+      // no config file
+    }
+
+    const customs = normalizeServiceConfig((config.llm as Record<string, unknown> | undefined)?.services)
+      .filter((s) => s.service === "custom")
+      .map((s) => ({
+        id: `custom:${s.name ?? "Custom"}`,
+        baseUrl: s.baseUrl ?? "",
+        label: s.name ?? "Custom",
+      }))
+      .filter((s) => s.baseUrl && Boolean(secrets.services[s.id]?.apiKey));
+
+    const groups = await Promise.all(customs.map(async (s) => ({
+      service: s.id,
+      label: s.label,
+      models: filterTextChatModels(
+        await probeModelsFromUpstream(s.baseUrl, secrets.services[s.id].apiKey, 10_000),
+      ),
+    })));
+
+    return c.json({ groups });
+  });
+
+  app.get("/api/v1/services/:service/models", async (c) => {
+    const service = c.req.param("service");
+    const refresh = c.req.query("refresh") === "1";
+    const secrets = await loadSecrets(options.root);
+    const apiKey = c.req.query("apiKey") || secrets.services[service]?.apiKey || "";
+
+    const resolvedBaseUrl = await resolveConfiguredServiceBaseUrl(options.root, service);
+    const baseService = isCustomServiceId(service) ? "custom" : service;
+    const apiKeyOptional = isApiKeyOptionalForEndpoint({
+      provider: resolveServiceProviderFamily(baseService) ?? "openai",
+      baseUrl: resolvedBaseUrl,
+    });
+
+    // No key = no models, except local/self-hosted endpoints such as Ollama.
+    if (!apiKey && !apiKeyOptional) return c.json({ models: [] });
+
+    // Cache by service + resolved baseUrl + apiKey fingerprint; valid for 10 min unless ?refresh=1
+    const cacheKey = `${service}::${resolvedBaseUrl ?? ""}::${apiKey.slice(-8)}`;
+    if (!refresh) {
+      const cached = modelListCache.get(cacheKey);
+      if (cached && Date.now() - cached.at < 10 * 60 * 1000) {
+        return c.json({ models: cached.models });
+      }
+    }
+
+    const enriched = await listModelsForService(
+      isCustomServiceId(service) ? "custom" : service,
+      apiKey,
+      isCustomServiceId(service) ? resolvedBaseUrl ?? undefined : undefined,
+    );
+    const models = filterTextChatModels(enriched).map((m) => ({
+      id: m.id,
+      name: m.name,
+      ...(m.maxOutput !== undefined ? { maxOutput: m.maxOutput } : {}),
+      ...(m.contextWindow > 0 ? { contextWindow: m.contextWindow } : {}),
+    }));
+    modelListCache.set(cacheKey, { models, at: Date.now() });
+    return c.json({ models });
+  });
+}
